CHAD-17085: zwave-button lazy loading of sub-drivers
