+++ .github/workflows/jenkins-driver-tests.yml
@@ -12,7 +12,7 @@ jobs:
     strategy:
       matrix:
         version:
-          [ 59, 60 ]
+          [ 59, 60, dev]
 
     runs-on: ubuntu-latest
     steps:
