+++ drivers/SmartThings/zwave-button/src/apiv6_bugfix/init.lua
@@ -1,13 +1,10 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
 local cc = require "st.zwave.CommandClass"
 local WakeUp = (require "st.zwave.CommandClass.WakeUp")({ version = 1 })
 
 
-local function can_handle(opts, driver, device, cmd, ...)
-  local version = require "version"
-  return version.api == 6 and
-    cmd.cmd_class == cc.WAKE_UP and
-    cmd.cmd_id == WakeUp.NOTIFICATION
-end
 
 local function wakeup_notification(driver, device, cmd)
   device:refresh()
@@ -20,7 +17,7 @@ local apiv6_bugfix = {
     }
   },
   NAME = "apiv6_bugfix",
-  can_handle = can_handle
+  can_handle = require("apiv6_bugfix.can_handle"),
 }
 
 return apiv6_bugfix
