+++ drivers/SmartThings/zwave-button/src/zwave-multi-button/aeotec-minimote/fingerprints.lua
@@ -1,0 +1,8 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local ZWAVE_AEOTEC_MINIMOTE_FINGERPRINTS = {
+  {mfr = 0x0086, prod = 0x0001, model = 0x0003} -- Aeotec Mimimote
+}
+
+return ZWAVE_AEOTEC_MINIMOTE_FINGERPRINTS
