+++ drivers/SmartThings/zwave-button/src/zwave-multi-button/init.lua
@@ -1,16 +1,7 @@
--- Copyright 2022 SmartThings
---
--- Licensed under the Apache License, Version 2.0 (the "License");
--- you may not use this file except in compliance with the License.
--- You may obtain a copy of the License at
---
---     http://www.apache.org/licenses/LICENSE-2.0
---
--- Unless required by applicable law or agreed to in writing, software
--- distributed under the License is distributed on an "AS IS" BASIS,
--- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
--- See the License for the specific language governing permissions and
--- limitations under the License.
+-- Copyright 2022 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+
 
 local capabilities = require "st.capabilities"
 --- @type st.zwave.CommandClass
@@ -20,33 +11,7 @@ local CentralScene = (require "st.zwave.CommandClass.CentralScene")({ version=1 
 --- @type st.zwave.CommandClass.SceneActivation
 local SceneActivation = (require "st.zwave.CommandClass.SceneActivation")({ version=1 })
 
-local ZWAVE_MULTI_BUTTON_FINGERPRINTS = {
-  {mfr = 0x010F, prod = 0x1001, model = 0x1000}, -- Fibaro KeyFob EU
-  {mfr = 0x010F, prod = 0x1001, model = 0x2000}, -- Fibaro KeyFob US
-  {mfr = 0x010F, prod = 0x1001, model = 0x3000}, -- Fibaro KeyFob AU
-  {mfr = 0x0371, prod = 0x0002, model = 0x0003}, -- Aeotec NanoMote Quad EU
-  {mfr = 0x0371, prod = 0x0102, model = 0x0003}, -- Aeotec NanoMote Quad US
-  {mfr = 0x0086, prod = 0x0001, model = 0x0058}, -- Aeotec KeyFob EU
-  {mfr = 0x0086, prod = 0x0101, model = 0x0058}, -- Aeotec KeyFob US
-  {mfr = 0x0086, prod = 0x0002, model = 0x0082}, -- Aeotec Wallmote Quad EU
-  {mfr = 0x0086, prod = 0x0102, model = 0x0082}, -- Aeotec Wallmote Quad US
-  {mfr = 0x0086, prod = 0x0002, model = 0x0081}, -- Aeotec Wallmote EU
-  {mfr = 0x0086, prod = 0x0102, model = 0x0081}, -- Aeotec Wallmote US
-  {mfr = 0x0060, prod = 0x000A, model = 0x0003}, -- Everspring Remote Control
-  {mfr = 0x0086, prod = 0x0001, model = 0x0003}, -- Aeotec Mimimote,
-  {mfr = 0x0371, prod = 0x0102, model = 0x0016}, -- Aeotec illumino Wallmote 7,
-  {mfr = 0x0460, prod = 0x0009, model = 0x0081}, -- Shelly Wave i4,
-  {mfr = 0x0460, prod = 0x0009, model = 0x0082}  -- Shelly Wave i4DC,
-}
 
-local function can_handle_zwave_multi_button(opts, driver, device, ...)
-  for _, fingerprint in ipairs(ZWAVE_MULTI_BUTTON_FINGERPRINTS) do
-    if device:id_match(fingerprint.mfr, fingerprint.prod, fingerprint.model) then
-      return true
-    end
-  end
-  return false
-end
 
 local map_key_attribute_to_capability = {
   [CentralScene.key_attributes.KEY_PRESSED_1_TIME] = capabilities.button.button.pushed,
@@ -115,12 +80,8 @@ local zwave_multi_button = {
   lifecycle_handlers = {
     init = device_init
   },
-  can_handle = can_handle_zwave_multi_button,
-  sub_drivers = {
-    require("zwave-multi-button/aeotec-keyfob"),
-    require("zwave-multi-button/fibaro-keyfob"),
-    require("zwave-multi-button/aeotec-minimote")
-  }
+  can_handle = require("zwave-multi-button.can_handle"),
+  sub_drivers = require("zwave-multi-button.sub_drivers"),
 }
 
 return zwave_multi_button
