+++ drivers/SmartThings/zwave-button/src/zwave-multi-button/shelly_wave_i4/fingerprints.lua
@@ -1,0 +1,9 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local SHELLY_WAVE_i4_FINGERPRINTS = {
+  {mfr = 0x0460, prod = 0x0009, model = 0x0081}, -- Shelly Wave i4
+  {mfr = 0x0460, prod = 0x0009, model = 0x0082}  -- Shelly Wave i4 DC
+}
+
+return SHELLY_WAVE_i4_FINGERPRINTS
